Programas/AWS/MTCars/requirements.txt
django
# instalado para MTCars
djangorestframework
mysqlclient==2.0.3
